fix(agent): use minute precision when rounded expiry is at least 60s in future (#1181)
In the previous behavior, assuming the current time is 10:00:00 and there's no clock drift:

- if a user provides `deltaInMs` equal to 85s, then rounded expiry is 10:01:25;
- if a user provides `deltaInMs` equal to 95s (i.e., larger than before), then rounded expiry is 10:01:00 (i.e., _earlier_ than before although the deltaInMs is _larger_).

With the new logic, we round the expiry to the nearest minute only if the rounding is at least 60 seconds in the future, otherwise we use second precision.

---

Thanks @mraszyk for spotting the bug and proposing the solution!