feat(agent): lookup canister ranges using the `/canister_ranges/<subnet_id>/<ranges>` certificate path (#1221)
# Description

Prepares the `check_canister_ranges` to support the certificate returned
by the new versions of the `call` and `read_state` APIs.

This is basically a porting of
https://github.com/dfinity/agent-rs/pull/672/files#diff-463e496396a1c352674f5d6dfe40ccb2527b1460c1e38662ecd3010743cda49cR977-R1000.

# How Has This Been Tested?

Since we are not supporting the new API versions, all the certificate
checks should fall into the `lookupCanisterRangesFallback` case, so all
existing tests should still work.